Implement higher-level tracking of VMAs
Prior to this commit, there was no high-level tracking of VMAs.
Everything was just stuffed into the page tables, and in order to
figure out of a page was mapped, one had to walk the hardware-level page
tables. This isn't great as it makes duplicating an address space during
a fork really difficult. Which pages correspond to the kernel? Which
ones correspond to userspace code?

This commit adds a linked list to each task_struct that tracks things at
a VMA level.