Implement ps2 driver and reading from stdin
This commit implements a PS2 keyboard driver and makes changes to the
scheduler to allow for reading from stdin (and sleeping if no input is
present). The PS2 driver is pretty straightforward, the scheduler
changes are nontrivial.

switch_to now switches to another process's stack in the kernel,
returning to wherever the last process was executing when it was
sleeping, allowing for easy waiting on IO. Newly created threads are
given a preformulated stack that returns to pop_regs_and_sysret (which
does exactly what it says).

When a process needs to wait on standard input, it removes itself from
the RR scheduler and calls schedule_sleep. The tty code will
subsequently wake the process up when enter is hit, and fill the buffer
passed to read().

init/shell zig binaries (which are pretty easy to compile for our
userspace) have been added to the zigprogs directory. The shell simply
prints a $ and attempts to exec (which currently fails).